feat(frontend): server-prefetch remaining detail pages
Job, Task, Experiment, Metric, Endpoint and project-scoped Endpoint
detail pages fetch their entity on the server and seed the client
component (local state or react-query initialData) from it. A 404 on
first load goes to the shared not-found page via notFoundIfEntityMissing;
later refetches keep the in-page missing-entity state.

useEndpoint/useProject accept an optional initialData. The
useIsCustomMetric hook is gone: it only re-fetched the metric to gate
the Tuning tab, which the prefetched metric now decides. Non-rhesis/
custom metrics redirect to /metrics on the server instead of via a
client toast.
