+++ apps/frontend/src/app/(protected)/experiments/[identifier]/page.tsx
@@ -3,7 +3,11 @@ export const dynamic = 'force-dynamic';
 import * as React from 'react';
 import { Alert, Paper } from '@mui/material';
 import { auth } from '@/auth';
-import ExperimentDetailClient from './components/ExperimentDetailClient';
+import { createServerApiFactory } from '@/utils/api-client/server-factory';
+import { notFoundIfEntityMissing } from '@/utils/entity-not-found-server';
+import ExperimentDetailClient, {
+  type ExperimentDetailData,
+} from './components/ExperimentDetailClient';
 
 interface PageProps {
   params: Promise<{ identifier: string }>;
@@ -23,5 +27,25 @@ export default async function ExperimentDetailPage({ params }: PageProps) {
   }
 
   const { identifier } = await params;
-  return <ExperimentDetailClient experimentId={identifier} />;
+  const client = (await createServerApiFactory()).getParametersClient();
+
+  let initialExperiment: ExperimentDetailData;
+  try {
+    const experiment = await client.getExperiment(identifier);
+    const [schema, environments] = await Promise.all([
+      client.getSchema(experiment.project_id),
+      client.getEnvironments(experiment.project_id),
+    ]);
+    initialExperiment = { experiment, schema, environments };
+  } catch (error) {
+    notFoundIfEntityMissing(error);
+    throw error;
+  }
+
+  return (
+    <ExperimentDetailClient
+      experimentId={identifier}
+      initialExperiment={JSON.parse(JSON.stringify(initialExperiment))}
+    />
+  );
 }
